+++ __tests__/UnitTest/InputTest.js
@@ -1,6 +1,6 @@
 import { Console } from '@woowacourse/mission-utils';
 import ERROR from '../../src/constants/error.js';
-import Input from '../../src/Input.js';
+import Input from '../../src/view/Input.js';
 
 const mockQuestions = (input) => {
   Console.readLineAsync = jest.fn();
