+++ src/constants/message.js
@@ -2,6 +2,7 @@ const MESSAGE = Object.freeze({
   PURCHASE_AMOUNT: '구입금액을 입력해 주세요.\n',
   WINNING_NUMBERS: '\n당첨 번호를 입력해 주세요.\n',
   BONUS_NUMBER: '\n보너스 번호를 입력해 주세요.\n',
+  BUYING_COUNT: (count) => `${count}개를 구매했습니다.\n`
 });
 
 export default MESSAGE;
