+++ src/view/Output.js
@@ -1,0 +1,3 @@
+class Output {
+
+}
